completed studio 1 exercises

+++ src/exercises/FuelEconomyCalculator.java
@@ -6,9 +6,9 @@ public class FuelEconomyCalculator {
     public static void main(String[] args) {
         Scanner input = new Scanner(System.in);
         System.out.println("How many miles have you driven?");
-        int miles = input.nextInt();
+        double miles = input.nextInt();
         System.out.println("How many gallons of gas have you consumed?");
-        int gallons = input.nextInt();
+        double gallons = input.nextInt();
         input.close();
         double mpg = miles / gallons;
         System.out.println("Your fuel economy is " + mpg + " miles per gallon.");
